etu_django_frame/django/4.2.16/requirements/requirement.txt
Django==4.2.16
celery==5.4.0
django-cas-ng==5.0.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-filter==24.3
django-redis-cache==3.0.1
django-redis-sessions==0.6.2
djangorestframework==3.15.2
django-simpleui==2024.8.28
ipython==8.14.0
kombu==5.4.0
vine==5.1.0
mysqlclient==2.2.0
redis==3.5.3
requests==2.32.3
uWSGI==2.0.26
etu-django-mcmt==1.0.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1